Move the types into a different file
